renamed the class script and init a script for addDepartment
