+++ lib/AddDepartment.js
@@ -1,27 +1,10 @@
-// Install Dependencies
-const inquirer = require('inquirer');
+const db = require('../config/connection.js');
 
-class AddDepartment {
-
-    constructor(){
-        this.name = "";
-    }
-
-    async setName(){
-        const response = await inquirer.prompt([
-            {
-                type: "input",
-                name: "text",
-                message: "Enter Department Name:",
-                validate: this.validateDepartmentName,
-            },
-        ]);
-        this.name = response.name;
-    }
-
-    async validateDepartmentName(input){
-        return (input.length >= 3 ) ? true : "Required minimum of 3 characters...";
-    }
+async function viewAllDepartments() {
+    db.query(`SELECT * FROM departments`, (err, data) => {
+        if (err) return console.log("Error reading database");
+        console.table(data);
+    });
 }
 
-module.exports = AddDepartment;
+module.exports =  viewAllDepartments ;
